refactor(pfb): extract preview-doc loading into a usePreviewDoc composable
