+++ frappe/public/js/print_format_builder/stores/index.js
@@ -1,6 +1,7 @@
 import { create_default_layout, serialize_layout } from "../utils";
 import { useLayoutHistory } from "./useLayoutHistory";
 import { usePresets } from "../composables/usePresets";
+import { usePreviewDoc } from "../composables/usePreviewDoc";
 import { watch, ref, inject, computed, nextTick } from "vue";
 
 // Copy/paste clipboard — persisted to localStorage so a field or section copied
@@ -76,11 +77,14 @@ export function getStore(print_format_name) {
 	let selected_section = ref(null);
 	let selected_letterhead = ref(false);
 	let selected_lh_footer = ref(false);
-	let preview_doc = ref(null);
-	let preview_doc_name = ref(null);
-	let preview_values = ref({});
-	let preview_child_values = ref({});
-	let preview_load_seq = 0;
+	const {
+		preview_doc,
+		preview_doc_name,
+		preview_values,
+		preview_child_values,
+		load_preview_doc,
+		persisted_preview_doc_name,
+	} = usePreviewDoc(print_format, print_format_name);
 
 	// methods
 	function fetch() {
@@ -278,38 +282,6 @@ export function getStore(print_format_name) {
 		) {
 			selected_field.value = null;
 		}
-	}
-	// Persist the chosen preview record per print format so it survives a refresh
-	const preview_doc_ls_key = `pfb:preview_doc:${print_format_name}`;
-	function persisted_preview_doc_name() {
-		return localStorage.getItem(preview_doc_ls_key);
-	}
-	function load_preview_doc(name) {
-		const seq = ++preview_load_seq;
-		if (!name) {
-			preview_doc.value = null;
-			preview_doc_name.value = null;
-			preview_values.value = {};
-			preview_child_values.value = {};
-			localStorage.removeItem(preview_doc_ls_key);
-			return;
-		}
-		preview_doc_name.value = name;
-		localStorage.setItem(preview_doc_ls_key, name);
-		frappe.db.get_doc(print_format.value.doc_type, name).then((doc) => {
-			if (seq !== preview_load_seq) return;
-			preview_doc.value = doc;
-		});
-		frappe
-			.call("frappe.utils.print_format_generator.get_formatted_field_values", {
-				doctype: print_format.value.doc_type,
-				name,
-			})
-			.then((r) => {
-				if (seq !== preview_load_seq) return;
-				preview_values.value = r.message?.values || {};
-				preview_child_values.value = r.message?.child || {};
-			});
 	}
 	function get_layout() {
 		if (print_format.value && print_format.value.format_data) {
